A api key can always revoke itself, add a route to delete any api key

+++ BTCPayServer/Security/GreenField/APIKeyRepository.cs
@@ -53,15 +53,18 @@ namespace BTCPayServer.Security.GreenField
             }
         }
 
-        public async Task Remove(string id, string getUserId)
+        public async Task<bool> Remove(string id, string getUserId)
         {
             using (var context = _applicationDbContextFactory.CreateContext())
             {
                 var key = await EntityFrameworkQueryableExtensions.SingleOrDefaultAsync(context.ApiKeys,
                     data => data.Id == id && data.UserId == getUserId);
+                if (key == null)
+                    return false;
                 context.ApiKeys.Remove(key);
                 await context.SaveChangesAsync();
             }
+            return true;
         }
 
         public class APIKeyQuery
